Moves and renames source actions
