+++ src/actions/sources/index.js
@@ -3,12 +3,10 @@
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
 // @flow
-import { clearSelectedSource } from "./clearSelectedSource";
-export { clearSelectedSource };
-
-export * from "./loadSourceText";
-export * from "./prettyPrint";
-export * from "./newSources";
 export * from "./blackbox";
+export * from "./sources";
+export * from "./loadSourceText";
+export * from "./newSources";
+export * from "./prettyPrint";
 export * from "./select";
 export * from "./tabs";
